==> Notebooks/04-Models.ipynb <==
## Modeling

Multiple machine learning models were used in this project, each with their unique capabilities to handle classification tasks. The models include Gradient Boosting, AdaBoost, Support Vector Classifier (SVC), Logistic Regression, Random Forest, and a Deep Neural Network.  

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, RandomizedSearchCV, GridSearchCV
from sklearn.metrics import ConfusionMatrixDisplay, classification_report, RocCurveDisplay
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import LinearSVC, SVC
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout, BatchNormalization
from tensorflow.keras.metrics import Recall
from tensorflow.keras.callbacks import EarlyStopping
from tensorflow.keras.regularizers import l2

credit = pd.read_csv('../data/train_cleaned.csv')
credit.head()

credit.columns

X = credit.drop(columns=[
    'id', 'is_delinquent', 'number_of_delinquent_months', 'average_delinquency_rate', 
    '3mo_delinquency', '6mo_delinquency', '12mo_delinquency'])

y = credit['is_delinquent']

cat_cols = X_train.select_dtypes(include='object').columns
num_cols = X_train.select_dtypes(include=['int64', 'float64']).columns

## Train-test(Validation) split

X = credit.drop(columns=[
    'id', 'is_delinquent', 'number_of_delinquent_months', 'average_delinquency_rate', 
    '3mo_delinquency', '6mo_delinquency', '12mo_delinquency'])

y_1 = credit['is_delinquent']
y_2 = credit['3mo_delinquency']
y_3 = credit['6mo_delinquency']
y_4 = credit['12mo_delinquency']


X_train, X_val, y_train_1, y_val_1 = train_test_split(X, y_1, test_size=0.2, random_state=42)

X_train, X_val, y_train_2, y_val_2 = train_test_split(X, y_2, test_size=0.2, random_state=42)
X_train, X_val, y_train_3, y_val_3 = train_test_split(X, y_3, test_size=0.2, random_state=42)

X_train, X_val, y_train_4, y_val_4 = train_test_split(X, y_4, test_size=0.2, random_state=42)

## One-Hot-Encoding and Standard Scaling variables with Column Transform

cat_cols = X_train.select_dtypes(include='object').columns
num_cols = X_train.select_dtypes(include=['int64', 'float64']).columns

sc = StandardScaler()

oh = OneHotEncoder(
    categories='auto', 
    drop='first', 
    sparse_output=False, 
    dtype='int',
    handle_unknown = 'ignore'
)

ct = ColumnTransformer(
    transformers=[
        ('oh', oh, cat_cols),
        ('sc', sc, num_cols)
    ]
)

X_train_ct = ct.fit_transform(X_train)
X_val_ct = ct.transform(X_val)

## Fit and scoring function

data_sets = [("is_delinquent", y_train_1, y_val_1), 
             ("3mo_delinquency", y_train_2, y_val_2), 
             ("6mo_delinquency", y_train_3, y_val_3)]

def fit_and_score(gs, X_train, y_train, X_val, y_val, name):
    print(f"\nPerforming grid search and scoring on dataset: {name}\n{'-'*50}")
    gs.fit(X_train, y_train)
    print(f'Training score on {name} : {gs.score(X_train, y_train)}')
    print(f'Testing score on {name} : {gs.score(X_val, y_val)}')
    best_params = gs.best_params_
    print(f"Best parameters for {name}:", best_params)
    return best_params


def fit_and_score_no_param(gs, X_train, y_train, X_val, y_val, name):
    print(f"\nPerforming grid search and scoring on dataset: {name}\n{'-'*50}")
    gs.fit(X_train, y_train)
    print(f'Training score on {name} : {gs.score(X_train, y_train)}')
    print(f'Testing score on {name} : {gs.score(X_val, y_val)}')
    return 



## Target Column Distributions

Predicting the classes with 3 different targets

print(credit['is_delinquent'].value_counts(normalize=True))
print(credit['3mo_delinquency'].value_counts(normalize=True))
print(credit['6mo_delinquency'].value_counts(normalize=True))

## Gradient Boosting with default parameters

gb = GradientBoostingClassifier()

for name, y_train, y_val in data_sets:
    fit_and_score_no_param(gb, X_train_ct, y_train, X_val_ct, y_val, name)


## Gradient Boosting with GridSearchCV

pgrids = {
    'learning_rate': [.15, .175, 0.2],
    'n_estimators': [285, 290, 295],
    'max_depth': [8, 9, 10]
}
    
gb = GradientBoostingClassifier()
gs = GridSearchCV(gb, param_grid=pgrids)

for name, y_train, y_val in data_sets:
    fit_and_score_no_param(gs, X_train_ct, y_train, X_val_ct, y_val, name)

## Ada Boost with GridSearchCV

%%time
tree = DecisionTreeClassifier(random_state=123)#, max_depth=1
ada = AdaBoostClassifier(estimator=tree, random_state=42)

pgrids_ada = {
    'learning_rate': [2.25, 2.5, 2.7],
    'n_estimators': [250, 300, 350],
    'estimator__max_depth': [71, 81, 91, 101, None],
    # 'estimator__min_samples_leaf': [7, 8, 9],
    'estimator__max_features': ['auto', 'sqrt', 'log2']
}


gs_ada_2 = GridSearchCV(ada, param_grid=pgrids_ada, cv=5)
gs_ada_2.fit(X_train_ct, y_train_3)

print(gs_ada.score(X_train_ct, y_train_3))
print(gs_ada.score(X_val_ct, y_val_3))  


## Support Vector Classifier

svc = SVC(C=1, kernel='rbf', degree = 1)

for name, y_train, y_val in data_sets:
    fit_and_score_no_param(svc, X_train_ct, y_train, X_val_ct, y_val, name)

## Logistic Regression

param_grid = {
    'C':  np.linspace(1, 10,5),
    'penalty': ['l2'],
    'solver': ['lbfgs'],
    'max_iter': [500,1000]
    
}

gs = GridSearchCV(
estimator= LogisticRegression(),
param_grid=param_grid)


for name, y_train, y_val in data_sets:
    fit_and_score(gs, X_train_ct, y_train, X_val_ct, y_val, name)

## Random Forest

import pickle

param_grid = {
    'n_estimators': [300,400],
    'max_depth': [None],
    'min_samples_split': [5,8],
    'min_samples_leaf': [3, 4],
    'max_features': ['sqrt']
}

rf = RandomForestClassifier()

gs = GridSearchCV(estimator=rf, param_grid=param_grid)

for name, y_train, y_val in data_sets:
    fit_and_score(gs, X_train_ct, y_train, X_val_ct, y_val, name)

    
with open("random_forest_model.pkl", "wb") as file:
    pickle.dump(gs, file)


## Dense Neural Network

def dnn(X_train, y_train, X_val, y_val):

    model = Sequential()
    model.add(Dense(256, input_dim=X_train_ct.shape[1], activation='relu'))
    model.add(Dense(128, activation='relu'))
    model.add(Dense(64, activation='relu'))
    model.add(Dense(32, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))

    model.compile(loss='binary_crossentropy', 
                  optimizer='adam', 
                  metrics=['accuracy', 'Recall'])

    es = EarlyStopping(monitor='val_loss', patience=5, restore_best_weights=True)

    model.fit(
    X_train_ct, 
    y_train_1, 
    epochs=50, 
    batch_size=32, 
    validation_data=(X_val, y_val),
    callbacks = es
    )

    return model

# Model did worse with dropouts, l2 regularizer
# Performed similar with extra hidden layers

is_delinq = dnn(X_train_ct, y_train_1, X_val_ct, y_val_1)

three_month_delinq = dnn(X_train_ct, y_train_2, X_val_ct, y_val_2)

six_month_delinq = dnn(X_train_ct, y_train_3, X_val_ct, y_val_3)
