==> Notebooks/05-Model-Evaluation.ipynb <==
# Assessing the Model Performance:

This notebook retrieves the best model along with its corresponding best set of hyperparameters that were trained in the 04 - Models section. The purpose of this step is to assess the performance of the model using unseen data.

By applying the best model to this new data, it is possible to evaluate how well it generalizes to examples it hasn't encountered during training. This assessment will provide insights into the model's ability to make accurate predictions and handle real-world scenarios.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import pickle

from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, RandomizedSearchCV, GridSearchCV
from sklearn.metrics import ConfusionMatrixDisplay, classification_report, RocCurveDisplay
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import LinearSVC, SVC
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from functions import credit_approval_data_cleaner, n_mo_delinquency

# import warnings
# warnings.filterwarnings("ignore")

applications_test = pd.read_csv('../data/applications_test.csv')
credits_test = pd.read_csv('../data/credits_test.csv')

test = credit_approval_data_cleaner(credits_test, applications_test, [6, 12])
test.head()

test.shape

### Importing and Setting up the Model

with open("random_forest_model.pkl", "rb") as file:
    best_model = pickle.load(file)

 best_model['model'].best_estimator_

## Model Evaluation

#### Importing and cleaning test data

X_test = test.drop(columns=[
    'id', 'is_delinquent', 'number_of_delinquent_months', 'average_delinquency_rate', 
    '6mo_delinquency', '12mo_delinquency'])

y_test = test['6mo_delinquency']

y_test.value_counts(normalize=True)

# print('Training set Accuracy:', bm.score(X, y))
# print('Test set Accuracy:', bm.score(X_test, y_test))
# print(f"Test set AUC:", roc_auc_score(y_test, bm.predict_proba(X_test)[:, 1]))


# print('\n')
# print(classification_report(y_test, bm.predict(X_test)))
# print('\n')

# fig, ax = plt.subplots(1, 2, figsize=(13, 6))
# ConfusionMatrixDisplay.from_estimator(bm, X_test, y_test, cmap='Blues', colorbar=False, ax=ax[0])
# RocCurveDisplay.from_estimator(bm, X_test, y_test, ax=ax[1])

# ax[0].set_title('Confusion Matrix')
# ax[1].set_title('ROC Curve')

# fig.suptitle('Model Performance', fontsize=18)

# plt.savefig('../images/model_performance.png', dpi=300);

### Conclusion:
